fix: renamed http and kafka to httpc and kafkac
