+++ connectors/httpc/source.go
@@ -1,4 +1,4 @@
-package source
+package httpc
 
 import (
 	"context"
